cacher: Implement a bare bones btree backed indexer
Wrap the cache.Indexer interface to extend it and provide
a method for performing clones. Clones in the btree impl
used follow a copy-on-write semantic.

For now, the implementation of only cache.Store is complete
along with the Clone() method. Remaining methods of the
cache.Indexer interface are stubbed out to satisfy interface
compliance.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>